+++ app/components/FullProfile/ProfileModal.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState, FC } from 'react';
-import { BrowserWallet, UTxO } from '@meshsdk/core';
+import { BrowserWallet, UTxO, MaestroProvider } from '@meshsdk/core';
 import { parseDatumCbor } from '@meshsdk/mesh-csl';
 import { hexToString } from '../utilities';
 import { CloseButton } from "./CloseButton";
@@ -9,6 +9,7 @@ import { ProfileForm } from './ProfileForm';
 import SuccessText from '../SuccessText';
 import Notification from '../Notification';
 import BlurImage from '../BlurImage';
+import { handleCreateCogno, handleDeleteCogno, handleUpdateCogno } from './transaction';
 
 interface ProfileModalProps {
   cogno: UTxO | null;
@@ -28,7 +29,87 @@ export const ProfileModal: FC<ProfileModalProps> = ({ cogno, network, wallet, on
   const [image, setImage] = useState('');
   const [details, setDetails] = useState('');
 
-  const clearNotification = () => setNotification('');
+  const clearNotification = () => {
+    setIsSuccessful(false);
+    setNotification('')
+  };
+
+  const checkTransaction = (network: number, message: string) => {
+    const networkName = network === 0 ? 'Preprod' : 'Mainnet';
+    const maestro = new MaestroProvider({ network: networkName, apiKey: process.env.NEXT_PUBLIC_MAESTRO!, turboSubmit: false });
+
+    const maxRetries = 250;
+
+    maestro.onTxConfirmed(message, async () => {
+      refreshCogno();
+      setNotification('transaction is on-chain');
+      setTitle('');
+      setImage('');
+      setDetails('');
+      setSubmittedTxHash('');
+      setIsSubmitting(false);
+    }, maxRetries);
+ 
+  };
+
+  const handleCreate = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setIsSubmitting(true);
+    setIsEditing(false);
+    setSubmittedTxHash('');
+    setIsSuccessful(false);
+    const result = await handleCreateCogno(network, wallet, { title, image, details});
+    if (result.success === false) {
+      // something failed so notify the user of the error message
+      setNotification(result.message);
+      setIsSubmitting(false);
+    } else {
+      // the transaction was submitted and we need to show the success modal
+      setSubmittedTxHash(result.message);
+      setIsSuccessful(true);
+      // this is where the actual sc interaction will be
+      checkTransaction(network!, result.message);
+    }
+  };
+
+  const handleUpdate = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setIsSubmitting(true);
+    setIsEditing(false);
+    setSubmittedTxHash('');
+    setIsSuccessful(false);
+    const result = await handleUpdateCogno(network, wallet, cogno, { title, image, details});
+    if (result.success === false) {
+      // something failed so notify the user of the error message
+      setNotification(result.message);
+      setIsSubmitting(false);
+    } else {
+      // the transaction was submitted and we need to show the success modal
+      setSubmittedTxHash(result.message);
+      setIsSuccessful(true);
+      // this is where the actual sc interaction will be
+      checkTransaction(network!, result.message);
+    }
+  };
+
+  const handleDelete = async () => {
+    setIsSubmitting(true);
+    setIsEditing(false);
+    setSubmittedTxHash('');
+    setIsSuccessful(false);
+    const result = await handleDeleteCogno(network, wallet, cogno);
+    if (result.success === false) {
+      // something failed so notify the user of the error message
+      setNotification(result.message);
+      setIsSubmitting(false);
+    } else {
+      // the transaction was submitted and we need to show the success modal
+      setSubmittedTxHash(result.message);
+      setIsSuccessful(true);
+      // this is where the actual sc interaction will be
+      checkTransaction(network!, result.message);
+    }
+  };
 
   useEffect(() => {
     if (cogno) {
@@ -45,23 +126,27 @@ export const ProfileModal: FC<ProfileModalProps> = ({ cogno, network, wallet, on
       <div className="fixed inset-0 dark-bg opacity-50 z-40"></div>
 
       {/* Modal content with a high z-index to ensure it's above everything */}
-      <div className="relative bg-white p-5 border rounded-md light-bg lg:max-w-screen-lg w-full z-50">
+      <div className="relative light-bg p-5 border rounded-md light-bg lg:max-w-screen-lg w-full z-50">
         <div className='flex items-center justify-between space-x-4'>
-          <h3 className="text-3xl font-medium dark-text">{title}</h3>
+          <h3 className="text-3xl font-medium dark-text text-nowrap text-ellipsis overflow-hidden">{title}</h3>
           {cogno && <Switch isOn={isEditing} onToggle={setIsEditing} />}
           {cogno ? (
             <ActionButton
               isSubmitting={!isEditing || isSubmitting}
-              onClick={() => console.log("Update")}
+              onClick={() => {}}
               className={`blue-bg ${isEditing ? 'blue-bg-hover' : ''} dark-text font-bold py-2 px-4 mx-1 rounded`}
+              type="submit"
+              form="profile-form"
             >
               Update
             </ActionButton>
           ) : (
             <ActionButton
               isSubmitting={isSubmitting}
-              onClick={() => console.log("Create")}
-              className="green-bg green-bg-hover dark-text font-bold py-2 px-4 mx-1 rounded"
+              onClick={() => {}}
+              className={`green-bg ${!isSubmitting ? 'green-bg-hover' : ''} dark-text font-bold py-2 px-4 mx-1 rounded`}
+              type="submit"
+              form="profile-form"
             >
               Create
             </ActionButton>
@@ -69,7 +154,7 @@ export const ProfileModal: FC<ProfileModalProps> = ({ cogno, network, wallet, on
           {cogno && (
             <ActionButton
               isSubmitting={!isEditing || isSubmitting}
-              onClick={() => console.log("Delete")}
+              onClick={handleDelete}
               className={`red-bg ${isEditing ? 'red-bg-hover' : ''} dark-text font-bold py-2 px-4 mx-1 rounded`}
             >
               Delete
@@ -88,7 +173,7 @@ export const ProfileModal: FC<ProfileModalProps> = ({ cogno, network, wallet, on
           )}
           <ProfileForm
             isSubmitting={cogno ? !isEditing || isSubmitting : isSubmitting}
-            handleSubmit={() => console.log("Submit Tx")}
+            handleSubmit={cogno ? handleUpdate : handleCreate}
             title={title}
             setTitle={setTitle}
             imageUrl={image}
@@ -107,7 +192,7 @@ export const ProfileModal: FC<ProfileModalProps> = ({ cogno, network, wallet, on
             Close
           </ActionButton>
         </div>
-        {notification && <Notification message={notification} onDismiss={clearNotification} />}
+        {notification && <Notification message={notification} onDismiss={clearNotification} successful={isSuccessful} />}
       </div>
     </div>
   );