+++ app/components/utilities.ts
@@ -6,3 +6,11 @@ export function hexToString(hex: string): string {
   }
   return str;
 }
+
+export function stringToHex(str: string): string {
+  let hex = '';
+  for (let i = 0; i < str.length; i++) {
+    hex += str.charCodeAt(i).toString(16).padStart(2, '0');
+  }
+  return hex;
+}